OPEN - task 1: Support for indexed value sets http://github.com/markus1978/emf-fragments/issues/issue/1
Implemented the grievances in this issue:
- refactored value set semantics (former value sets)
- moved from user site list implementation to internal setting delegate
based list
- added tests for value sets with containment and references
- implemented indexed list class, but not tested yet
- realized that tests must be improved
- many list operations are not supported

+++ de.hub.emffrag/src/de/hub/emffrag/model/emffrag/IndexedList.java
@@ -25,10 +25,10 @@ import org.eclipse.emf.ecore.EObject;
  *
  *
  * @see de.hub.emffrag.model.emffrag.EmfFragPackage#getIndexedList()
- * @model superTypes="de.hub.emffrag.model.emffrag.IndexedMap<org.eclipse.emf.ecore.EIntegerObject, V>"
+ * @model superTypes="de.hub.emffrag.model.emffrag.IndexedMap<org.eclipse.emf.ecore.ELongObject, V>"
  * @generated
  */
-public interface IndexedList<V extends EObject> extends IndexedMap<Integer, V> {
+public interface IndexedList<V extends EObject> extends IndexedMap<Long, V> {
 	/**
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
